Fix tests nad move them to content
