+++ cypress/integration/Content/content.ts
@@ -1,6 +1,17 @@
 /// <reference types="Cypress" />
-import { DocumentTypeBuilder, ContentBuilder, AliasHelper } from '../../../src';
+import {
+    DocumentTypeBuilder,
+    ContentBuilder,
+    AliasHelper,
+    GridDataTypeBuilder,
+    PartialViewMacroBuilder, MacroBuilder
+} from '../../../src';
+
 context('Content', () => {
+
+    beforeEach(() => {
+        cy.umbracoLogin(Cypress.env('username'), Cypress.env('password'));
+    });
 
     function refreshContentTree(){
         // Refresh to update the tree
@@ -10,9 +21,22 @@ context('Content', () => {
         cy.get('.umb-tree-item__inner').should('exist', {timeout: 10000});
     }
 
-    beforeEach(() => {
-        cy.umbracoLogin(Cypress.env('username'), Cypress.env('password'));
-    });
+    function createSimpleMacro(name){
+        const insertMacro = new PartialViewMacroBuilder()
+            .withName(name)
+            .withContent(`@inherits Umbraco.Web.Macros.PartialViewMacroPage
+<h1>Acceptance test</h1>`)
+            .build();
+
+        const macroWithPartial = new MacroBuilder()
+            .withName(name)
+            .withPartialViewMacro(insertMacro)
+            .withRenderInEditor()
+            .withUseInEditor()
+            .build();
+
+        cy.saveMacroWithPartial(macroWithPartial);
+    }
 
     it('Copy content', () => {
         const rootDocTypeName = "Test document type";
@@ -82,7 +106,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Copy node
         cy.umbracoTreeItem("content", [nodeName, childNodeName]).rightclick({ force: true });
@@ -166,7 +190,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Move node
         cy.umbracoTreeItem("content", [nodeName, childNodeName]).rightclick({ force: true });
@@ -253,7 +277,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Sort nodes
         cy.umbracoTreeItem("content", [nodeName]).rightclick({ force: true });
@@ -302,7 +326,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Access node
         cy.umbracoTreeItem("content", [initialNodeName]).click();
@@ -313,15 +337,16 @@ context('Content', () => {
 
         // Save and publish
         cy.get('.btn-success').first().click();
+        cy.umbracoSuccessNotification().should('be.visible');
 
         // Rollback
         cy.get('.umb-box-header :button').click();
 
-        cy.get('.umb-box-content > .ng-scope > .input-block-level')
+        cy.get('.umb-box-content > div > .input-block-level')
             .find('option[label*=' + new Date().getDate() + ']')
             .then(elements => {
-                const option = elements[[elements.length - 1]].getAttribute('value');
-                cy.get('.umb-box-content > .ng-scope > .input-block-level')
+                const option = elements[elements.length - 1].getAttribute('value');
+                cy.get('.umb-box-content > div > .input-block-level')
                     .select(option);
             });
 
@@ -330,8 +355,8 @@ context('Content', () => {
         cy.reload();
 
         // Assert
-        cy.get('.history').find('.umb-badge').eq(0).should('contain.text', "Save");
-        cy.get('.history').find('.umb-badge').eq(1).should('contain.text', "Rollback");
+        cy.get('.history').find('.umb-badge').contains('Save').should('be.visible');
+        cy.get('.history').find('.umb-badge').contains('Rollback').should('be.visible');
         cy.get('#headerName').should('have.value', initialNodeName);
 
         // Clean up (content is automatically deleted when document types are gone)
@@ -369,7 +394,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Access node
         cy.umbracoTreeItem("content", [nodeName]).click();
@@ -410,7 +435,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Access node
         cy.umbracoTreeItem("content", [nodeName]).click();
@@ -448,7 +473,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Access node
         cy.umbracoTreeItem("content", [nodeName]).click();
@@ -488,7 +513,7 @@ context('Content', () => {
         });
 
         // Refresh to update the tree
-        refreshContentTree()
+        refreshContentTree();
 
         // Access node
         cy.umbracoTreeItem("content", [nodeName]).click();
@@ -501,11 +526,9 @@ context('Content', () => {
     });
 
     it('Content with contentpicker', () => {
-        const pickerDocTypeName = 'Content Picker Type';
-        const groupName = 'ContentPickerGroup';
-        const alias = AliasHelper.toAlias(pickerDocTypeName);
-        const pickedDocTypeName = 'Picked Content Type';
-        const pickedNodeName = 'Content to pick';
+        const pickerDocTypeName = 'Content picker doc type';
+        const pickerDocTypeAlias = AliasHelper.toAlias(pickerDocTypeName);
+        const pickedDocTypeName = 'Picked content document type';
 
         cy.deleteAllContent();
         cy.umbracoEnsureDocumentTypeNameNotExists(pickerDocTypeName);
@@ -522,13 +545,13 @@ context('Content', () => {
             .done()
             .done()
             .build();
-            
+
         cy.saveDocumentType(pickedDocType).then((generatedType) => {
             const pickedContentNode = new ContentBuilder()
                 .withContentTypeAlias(generatedType["alias"])
                 .withAction("publishNew")
                 .addVariant()
-                .withName(pickedNodeName)
+                .withName('Content to pick')
                 .withSave(true)
                 .withPublish(true)
                 .addProperty()
@@ -545,11 +568,11 @@ context('Content', () => {
         // Create the doctype with a the picker
         const pickerDocType = new DocumentTypeBuilder()
             .withName(pickerDocTypeName)
-            .withAlias(alias)
+            .withAlias(pickerDocTypeAlias)
             .withAllowAsRoot(true)
-            .withDefaultTemplate(alias)
+            .withDefaultTemplate(pickerDocTypeAlias)
             .addGroup()
-            .withName(groupName)
+            .withName('ContentPickerGroup')
             .addContentPickerProperty()
             .withAlias('picker')
             .done()
@@ -558,13 +581,13 @@ context('Content', () => {
 
         cy.saveDocumentType(pickerDocType);
 
-        // Edit the template to allow us to verify the rendered view.
-        cy.editTemplate(pickerDocTypeName, `@inherits Umbraco.Web.Mvc.UmbracoViewPage<ContentModels.ContentPickerType>
+        // Edit it the template to allow us to verify the rendered view.
+        cy.editTemplate(pickerDocTypeName, `@inherits Umbraco.Web.Mvc.UmbracoViewPage<ContentModels.ContentPickerDocType>
         @using ContentModels = Umbraco.Web.PublishedModels;
         @{
             Layout = null;
         }
-        
+
         @{
             IPublishedContent typedContentPicker = Model.Value<IPublishedContent>("picker");
             if (typedContentPicker != null)
@@ -575,8 +598,8 @@ context('Content', () => {
 
         // Create content with content picker
         cy.get('.umb-tree-root-link').rightclick();
-        cy.get('.-opens-dialog > .umb-action-link').click();
-        cy.get('[data-element="action-create-contentPickerType"] > .umb-action-link').click();
+        cy.get('[data-element="action-create"]').click();
+        cy.get('[data-element="action-create-' + pickerDocTypeAlias + '"] > .umb-action-link').click();
         // Fill out content
         cy.umbracoEditorHeaderName('ContentPickerContent');
         cy.get('.umb-node-preview-add').click();
@@ -589,11 +612,195 @@ context('Content', () => {
         cy.umbracoSuccessNotification().should('be.visible');
 
         // Assert
+        cy.log('Checking that content is rendered correctly.')
         const expectedContent = '<p>Acceptance test</p>'
         cy.umbracoVerifyRenderedViewContent('contentpickercontent', expectedContent, true).should('be.true');
         // clean
         cy.umbracoEnsureDocumentTypeNameNotExists(pickerDocTypeName);
         cy.umbracoEnsureTemplateNameNotExists(pickerDocTypeName);
         cy.umbracoEnsureDocumentTypeNameNotExists(pickedDocTypeName);
+    });
+
+    it('Content with macro in RTE', () => {
+        const viewMacroName = 'Content with macro in RTE';
+        const partialFileName = viewMacroName + '.cshtml';
+
+        cy.umbracoEnsureMacroNameNotExists(viewMacroName);
+        cy.umbracoEnsurePartialViewMacroFileNameNotExists(partialFileName);
+        cy.umbracoEnsureDocumentTypeNameNotExists(viewMacroName);
+        cy.umbracoEnsureTemplateNameNotExists(viewMacroName);
+        cy.deleteAllContent();
+
+        // First thing first we got to create the macro we will be inserting
+        createSimpleMacro(viewMacroName);
+
+        // Now we need to create a document type with a rich text editor where we can insert the macro
+        // The document type must have a template as well in order to ensure that the content is displayed correctly
+        const alias = AliasHelper.toAlias(viewMacroName);
+        const docType = new DocumentTypeBuilder()
+            .withName(viewMacroName)
+            .withAlias(alias)
+            .withAllowAsRoot(true)
+            .withDefaultTemplate(alias)
+            .addGroup()
+            .addRichTextProperty()
+            .withAlias('text')
+            .done()
+            .done()
+            .build();
+
+        cy.saveDocumentType(docType).then((generatedDocType) => {
+            // Might as wel initally create the content here, the less GUI work during the test the better
+            const contentNode = new ContentBuilder()
+                .withContentTypeAlias(generatedDocType["alias"])
+                .withAction('saveNew')
+                .addVariant()
+                .withName(viewMacroName)
+                .withSave(true)
+                .done()
+                .build();
+
+            cy.saveContent(contentNode);
+        });
+
+        // Edit the macro template in order to have something to verify on when rendered.
+        cy.editTemplate(viewMacroName, `@inherits Umbraco.Web.Mvc.UmbracoViewPage
+@using ContentModels = Umbraco.Web.PublishedModels;
+@{
+  Layout = null;
+}
+@{
+    if (Model.HasValue("text")){
+        @(Model.Value("text"))
+    }
+} `);
+
+        // Enter content
+        refreshContentTree();
+        cy.umbracoTreeItem("content", [viewMacroName]).click();
+
+        // Insert macro
+        cy.get('#mceu_13-button').click();
+        cy.get('.umb-card-grid-item').contains(viewMacroName).click();
+
+        // Save and publish
+        cy.umbracoButtonByLabelKey('buttons_saveAndPublish').click();
+        cy.umbracoSuccessNotification().should('be.visible');
+
+        // Ensure that the view gets rendered correctly
+        const expected = `<h1>Acceptance test</h1><p> </p>`;
+        cy.umbracoVerifyRenderedViewContent('/', expected, true).should('be.true');
+
+        // Cleanup
+        cy.umbracoEnsureMacroNameNotExists(viewMacroName);
+        cy.umbracoEnsurePartialViewMacroFileNameNotExists(partialFileName);
+        cy.umbracoEnsureDocumentTypeNameNotExists(viewMacroName);
+        cy.umbracoEnsureTemplateNameNotExists(viewMacroName);
+    });
+
+    it('Content with macro in grid', () => {
+        const name = 'Content with macro in grid';
+        const macroName = 'Grid macro';
+        const macroFileName = macroName + '.cshtml';
+
+        cy.umbracoEnsureDataTypeNameNotExists(name);
+        cy.umbracoEnsureDocumentTypeNameNotExists(name);
+        cy.umbracoEnsureTemplateNameNotExists(name);
+        cy.umbracoEnsureMacroNameNotExists(macroName);
+        cy.umbracoEnsurePartialViewMacroFileNameNotExists(macroFileName);
+        cy.deleteAllContent();
+
+        createSimpleMacro(macroName);
+
+        const grid = new GridDataTypeBuilder()
+            .withName(name)
+            .withDefaultGrid()
+            .build();
+
+        const alias = AliasHelper.toAlias(name);
+        // Save grid and get the ID 
+        cy.saveDataType(grid).then((dataType) => {
+            // Create a document type using the data type
+            const docType = new DocumentTypeBuilder()
+                .withName(name)
+                .withAlias(alias)
+                .withAllowAsRoot(true)
+                .withDefaultTemplate(alias)
+                .addGroup()
+                .addCustomProperty(dataType['id'])
+                .withAlias('grid')
+                .done()
+                .done()
+                .build();
+
+            cy.saveDocumentType(docType).then((generatedDocType) => {
+                const contentNode = new ContentBuilder()
+                    .withContentTypeAlias(generatedDocType["alias"])
+                    .addVariant()
+                    .withName(name)
+                    .withSave(true)
+                    .done()
+                    .build();
+
+                cy.saveContent(contentNode);
+            });
+        });
+
+        // Edit the template to allow us to verify the rendered view
+        cy.editTemplate(name, `@inherits Umbraco.Web.Mvc.UmbracoViewPage
+@using ContentModels = Umbraco.Web.PublishedModels;
+@{
+  Layout = null;
+}
+@Html.GetGridHtml(Model, "grid")`);
+
+        // Act
+        // Enter content
+        refreshContentTree();
+        cy.umbracoTreeItem("content", [name]).click();
+        // Click add
+        cy.get(':nth-child(2) > .preview-row > .preview-col > .preview-cell').click(); // Choose 1 column layout.
+        cy.get('.umb-column > .templates-preview > :nth-child(2) > .ng-binding').click(); // Choose headline
+        cy.get('.umb-cell-placeholder').click();
+        // Click macro
+        cy.get(':nth-child(4) > .umb-card-grid-item > :nth-child(1)').click();
+        // Select the macro
+        cy.get('.umb-card-grid-item').contains(macroName).click();
+
+        // Save and publish
+        cy.umbracoButtonByLabelKey('buttons_saveAndPublish').click();
+        cy.umbracoSuccessNotification().should('be.visible');
+
+        const expected = `
+    <div class="umb-grid">
+                <div class="grid-section">
+    <div >
+            <div class="container">
+        <div class="row clearfix">
+            <div class="col-md-12 column">
+                <div >
+
+
+
+
+        <h1>Acceptance test</h1>
+
+
+
+                </div>
+            </div>        </div>
+            </div>
+    </div>
+                </div>
+    </div>`
+
+        cy.umbracoVerifyRenderedViewContent('/', expected, true).should('be.true');
+
+        // Clean 
+        cy.umbracoEnsureDataTypeNameNotExists(name);
+        cy.umbracoEnsureDocumentTypeNameNotExists(name);
+        cy.umbracoEnsureTemplateNameNotExists(name);
+        cy.umbracoEnsureMacroNameNotExists(macroName);
+        cy.umbracoEnsurePartialViewMacroFileNameNotExists(macroFileName);
     });
 });
