+++ package.json
@@ -51,7 +51,7 @@
     "@typescript-eslint/parser": "^3.9.0",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
-    "cypress": "^5.1.0",
+    "cypress": "^6.8.0",
     "cypress-file-upload": "^4.0.7",
     "cypress-wait-until": "^1.7.1",
     "eslint": "^7.6.0",
