+++ src/cms/builders/dataTypes/gridBuilders/index.ts
@@ -2,7 +2,7 @@ export { GridLayoutBuilder } from './gridLayoutBuilder';
 export { SectionBuilder } from './sectionBuilder';
 export { GridRowConfigBuilder } from './gridRowConfigBuilder';
 export { GridAreaBuilder } from './gridAreaBuilder';
-export { GridRteBuilder } from './gridRteBuilder'
+export { GridRteBuilder } from './gridRteBuilder';
 export { RteToolbarOptionsBuilder } from './rteToolbarOptionsBuilder';
 export { GridSettingsbuilder } from './gridSettingsBuilder';
 export { GridStylesBuilder } from './gridStylesBuilder';
